Moved core_table_filter_utils to the utils folder
