+++ frontend/src/store/utils/core_table_filter_utils.ts
@@ -16,8 +16,8 @@ limitations under the License. */
 //  they can be mocked by Jest:
 //  https://stackoverflow.com/questions/51269431/jest-mock-inner-function
 
-import { RecommendationExtra } from "./recommendation_extra";
-import { ICoreTableStoreState, costCategoriesNames } from "./core_table";
+import { RecommendationExtra } from "../recommendation_extra";
+import { ICoreTableStoreState, costCategoriesNames } from "../core_table";
 
 // case-insensitive search (searching for 'myvm' should match 'MyVm')
 export function isSearchTextInCell(
